requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1600297518631/work
argon2-cffi @ file:///home/conda/feedstock_root/build_artifacts/argon2-cffi_1596629847793/work
astunparse==1.6.3
async-generator==1.10
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1599308529326/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache==1.6.1
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1600454382015/work
blinker==1.4
brotlipy==0.7.0
bsuite @ git+git://github.com/deepmind/bsuite.git@f2e6c2103247579498e2b84ca7581d8ae4632dac
cachetools @ file:///tmp/build/80754af9/cachetools_1596822027882/work
certifi==2020.6.20
cffi @ file:///tmp/build/80754af9/cffi_1598370769933/work
chardet==3.0.4
chex==0.0.2
click==7.1.2
cloudpickle==1.3.0
coloredlogs==14.0
crowdai-api==0.1.22
cryptography @ file:///tmp/build/80754af9/cryptography_1598892038851/work
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
descartes==1.1.0
dill==0.3.2
-e git+https://github.com/dzorlu/acme.git@880a001187dd4edb32f608a1b7e445b0f9899e71#egg=dm_acme
dm-control==0.0.322773188
dm-env==1.2
dm-haiku==0.0.2
dm-reverb-nightly==0.1.0.dev20200708
dm-sonnet==2.0.0
dm-tree==0.1.5
entrypoints==0.3
frozendict==1.2
future==0.18.2
gast==0.3.3
getch==1.0
glfw==1.12.0
google-auth @ file:///tmp/build/80754af9/google-auth_1600274525154/work
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio @ file:///tmp/build/80754af9/grpcio_1597424474635/work
gym==0.17.2
h5py @ file:///tmp/build/80754af9/h5py_1593454122442/work
humanfriendly==8.2
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.9.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1593446406207/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1595446881062/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1598749948871/work
ipython-genutils==0.2.0
jax==0.1.77
jaxlib==0.1.55
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
json5 @ file:///home/conda/feedstock_root/build_artifacts/json5_1591810480056/work
jsonschema==3.2.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1598486169312/work
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.1
jupyterlab-server @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_server_1593951277307/work
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
labmaze==1.0.3
lxml==4.5.2
Markdown @ file:///tmp/build/80754af9/markdown_1597433240441/work
MarkupSafe==1.1.1
matplotlib==3.0.3
minerl==0.3.6
mistune==0.8.4
mizani==0.7.1
mkl-fft==1.1.0
mkl-random==1.1.0
mkl-service==2.3.0
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1598558657104/work
nbconvert @ file:///home/conda/feedstock_root/build_artifacts/nbconvert_1600286912556/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1594060262917/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1594996608835/work
networkx==2.5
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1599742234243/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1596233721170/work
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.1.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1589925210001/work
palettable==3.3.0
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotnine==0.7.1
portpicker==1.3.1
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1590412252446/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1598885455507/work
protobuf==3.12.4
psutil==5.7.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyglet==1.5.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1600347314331/work
PyJWT==1.7.1
PyOpenGL==3.1.5
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
Pyro4==4.80
pyrsistent @ file:///home/conda/feedstock_root/build_artifacts/pyrsistent_1599988181994/work
PySocks==1.7.1
python-dateutil==2.8.1
python-gitlab==2.5.0
pytz==2020.1
PyWavelets==1.1.1
pyzmq==19.0.2
redis==3.5.3
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rlax @ git+git://github.com/deepmind/rlax.git@5e1a6bfd7271a150bc5c2568f7a2f7648863fc04
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.4.1
Send2Trash==1.5.0
serpent==1.30.2
six==1.15.0
sklearn==0.0
statsmodels==0.12.0
tabulate==0.8.7
tb-nightly==2.3.0a20200722
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==2.4.0.dev2020091801
tf-nightly==2.4.0.dev20200708
tfp-nightly==0.12.0.dev20200717
threadpoolctl==2.1.0
tifffile==2020.9.3
toolz==0.10.0
tornado==6.0.4
tqdm==4.49.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1599471676085/work
trfl==1.1.0
typing==3.7.4.3
urllib3 @ file:///tmp/build/80754af9/urllib3_1597086586889/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1595859607677/work
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
